Make setUlimit a noop on Windows.
