+++ ibazel/main.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"os"
 	"strings"
-	"syscall"
 	"time"
 )
 
@@ -162,25 +161,4 @@ func handle(i *IBazel, command string, args []string) {
 		usage()
 		return
 	}
-}
-
-func setUlimit() error {
-	var lim syscall.Rlimit
-
-	err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &lim)
-	if err != nil {
-		return err
-	}
-
-	// set the "soft" file descriptor to the maximum
-	// allowed by a userspace program.
-	// http://man7.org/linux/man-pages/man2/getrlimit.2.html
-	lim.Cur = lim.Max
-
-	err = syscall.Setrlimit(syscall.RLIMIT_NOFILE, &lim)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
